requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==2.2
django-decouple==2.1
djangorestframework==3.12.2
mysqlclient==1.4.6
pkg-resources==0.0.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
